requirements.txt
customtkinter==5.1.2
darkdetect==0.8.0
Pillow==9.4.0
typing-extensions==4.5.0
